% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypotenuse_length.R
\name{hypotenuse_length}
\alias{hypotenuse_length}
\title{Hypotenuse Length for Right Angle Triangle}
\usage{
hypotenuse_length(a = 1, b = 1)
}
\arguments{
\item{a}{This is the length of first right-angle side}

\item{b}{This is the length of second right-angle side}
}
\value{
The function returns the length of hypotenuse in the right angle triangle
}
\description{
Hypotenuse Length for Right Angle Triangle
}
\examples{
hypotenuse_length(2,2)
}
