% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_binomial_sample.R
\name{mean_binomial_sample}
\alias{mean_binomial_sample}
\title{Mean of Binomial Samples}
\usage{
mean_binomial_sample(samplesize, n = 2, p = 0.5)
}
\arguments{
\item{samplesize}{This is the size of the random sample that you want to generate}

\item{n}{This is the number of trials in the binomial distribution}

\item{p}{This is the probability in the binomial distribution}
}
\value{
The function returns the mean of the random sample of binomial distribution generated.
}
\description{
Mean of Binomial Samples
}
\examples{
mean_binomial_sample(50,10,0.5)
}
